pyproject.toml
[project]
name = "vlada-dashboard"
version = "0.1.0"
description = "Streamlit Application for Wedding Info"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pymongo>=4.11.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.43.2",
    "streamlit-plotly-events>=0.0.6",
]
